added the Mandelbrot to ImGui
