+++ imgui.ini
@@ -3,8 +3,9 @@ Pos=60,60
 Size=400,400
 
 [Window][Dear ImGui Demo]
-Pos=192,76
+Pos=105,11
 Size=550,680
+Collapsed=1
 
 [Window][Hello, world!]
 Pos=31,71
